requirements/base.txt
# Core dependencies
#numpy
#pandas
#matplotlib
#keras
#tensorflow
#pydot
#pydotplus
#graphviz

# Projet V1 - BASE
jinja2
fastapi
uvicorn
python-multipart
pandas
matplotlib
keras
tensorflow
pillow
numpy
httpx
requests
python-dotenv

# Projet V2 - MONITORING
sqlalchemy
psycopg2-binary
pydantic
plotly
pytest

# Projet V3 - MLOPS
